requirements.txt
fastapi
uvicorn
requests
PyMuPDF
python-docx
faiss-cpu
python-dotenv
google-generativeai
